backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
requests==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
langchain==0.0.352
openai==1.3.5
qdrant-client==1.7.0
better-auth==0.4.3